requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
aiohttp==3.10.2
pillow==10.4.0
python-dotenv==1.0.1
